Text comments fixed File name corrected
