+++ app/chains/__init__.py
@@ -2,6 +2,6 @@ BASE_ROUTE = 'chains'
 
 
 def register_routes(api, app, root="api"):
-    from .models import api as chains_api
+    from .controller import api as chains_api
 
     api.add_namespace(chains_api, path=f"/{root}/{BASE_ROUTE}")
