--- app/chains/chain.py
@@ -1,13 +1,0 @@
-from dataclasses import dataclass
-
-
-@dataclass
-class Chain:
-    uid: str
-    nodes: dict
-
-
-@dataclass
-class ChainResponse:
-    uid: str
-    is_new: bool
